+++ www/js/wikidiverController.ts
@@ -13,6 +13,10 @@ declare var pageSearchResultDetail: any;
 /*
 TODO
 
+  2016/07/15
+    statusを表示したい
+    データ操作後の処理をしっかりしたい
+
   2016/07/12
     残りタスク
       image表示, 表示用のオプション追加
@@ -76,7 +80,7 @@ TODO
       let m_lang_ap = storage_manager_settings.getItem(SETTING_TYPE.LANGUAGE_APPEARANCE);
       if(!m_lang_ap){
         m_lang_ap = "ja";
-        storage_manager_settings.saveItem2Storage(SETTING_TYPE.LANGUAGE, m_lang_ap);
+        storage_manager_settings.saveItem2Storage(SETTING_TYPE.LANGUAGE_APPEARANCE, m_lang_ap);
       }
       // 取得先言語設定
       let m_lang = storage_manager_settings.getItem(SETTING_TYPE.LANGUAGE);
@@ -101,6 +105,12 @@ TODO
       if(!m_his_len){
         m_his_len = 10;
         storage_manager_settings.saveItem2Storage(SETTING_TYPE.HISTORY_LENGTH, m_his_len);
+      }
+
+      // global memoが存在しない場合、空白をセットしておく
+      let g_memo = storage_manager_memo.getItem(GLOBAL_MEMO_PROP.KEY);
+      if(!g_memo){
+        storage_manager_memo.saveItem2Storage(GLOBAL_MEMO_PROP.KEY, "");
       }
 
 
@@ -160,6 +170,7 @@ TODO
       };
 
       $scope.move2setting = function(){
+        myMenu.closeMenu();
         myNavigator.pushPage("settings.html");
       };
 
@@ -169,7 +180,7 @@ TODO
     module.controller("HomeController", function($scope, popoverSharingService){
         $scope.search_key = "";
         $scope.favorite_length = storage_manager_favorite.getItemLength();
-        $scope.history_length = storage_manager_history.getItemLength();
+        $scope.history_length = storage_manager_history.getItemLength() - 1; // 常にglobalMemoが存在する仕様なので
         $scope.notes_length = storage_manager_memo.getItemLength();
 
 
@@ -210,9 +221,9 @@ TODO
 
           // popover オープン前に必要情報をコピー
           popoverSharingService.sharing.id = GLOBAL_MEMO_PROP.KEY;
-          popoverSharingService.sharing.title = $scope.title;
+          popoverSharingService.sharing.title = GLOBAL_MEMO_PROP.KEY;
           popoverSharingService.sharing.caption = "";
-          popoverSharingService.sharing.memo = g_memo || "";
+          popoverSharingService.sharing.memo = g_memo.memo || "";
           popoverSharingService.sharing.is_global = true; // globalメモなので
 
           // sharingの値更新をsubscribeする
@@ -821,12 +832,22 @@ $scope._showImages = function(){
         popoverSharingService.sharing = $scope.sharing; // 次ポップアップオープン時用にコピー
 
         showAlert("save success!! but, actually needed to check success or failure...");
+
+        myPopoverMemo.hide();
       };
 
       $scope.deleteMemo = function(){
-        storage_manager_memo.deleteItem($scope.sharing.id);
+        // global memoの場合は、本当に削除はしない
+        if($scope.id != GLOBAL_MEMO_PROP.KEY){
+          storage_manager_memo.deleteItem($scope.sharing.id);
+        }
+        else{
+          storage_manager_memo.saveItem2Storage(GLOBAL_MEMO_PROP.KEY, "");
+        }
         $scope.sharing.memo = "";
         showAlert("delete success!! but, actually needed to check success or failure...");
+
+        myPopoverMemo.hide();
       };
 
       // close押下時
@@ -942,7 +963,7 @@ $scope._showImages = function(){
           $scope.history_length = 0;
         }
 
-        storage_manager_settings.saveItem2Storage(SETTING_TYPE.LANGUAGE_APPEARANCE, $scope.radio_language);
+        storage_manager_settings.saveItem2Storage(SETTING_TYPE.LANGUAGE_APPEARANCE, $scope.radio_language_appearance);
         storage_manager_settings.saveItem2Storage(SETTING_TYPE.LANGUAGE, $scope.radio_language);
         storage_manager_settings.saveItem2Storage(SETTING_TYPE.IMG_HANDLE, $scope.radio_imghandle);
         storage_manager_settings.saveItem2Storage(SETTING_TYPE.ARTICLE_TYPE, $scope.radio_article);
